Add biflowDirection field to IPFIX report
Add the biflowDirection field to the IPFIX report. We derive
the field settings from whoever sent the first SYN packet.
If we don't observe a SYN packet, we leave this field at 0, which
means arbitrary assignment. Therefore, we will only be able to
see values indicating that source or destination initiated the
flow when the first packet was observed (TCP).

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>